# Construct fallback slice call with symbols, not values